requirements.txt
djangorestframework
djangorestframework-simplejwt
markdown
django-filter